back/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
stripe
pydantic
